# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"messageDateResponse": {
				"properties": {
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"date",
					"message"
				],
				"type": "object",
				"description": "MessageDateResponse json response controller"
			},
			"BasicResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"description": "Basic Json Response Controller"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "code-verifier",
		"version": "1.0.0",
		"description": "Backed Node + TS + Swagger",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/bye": {
			"get": {
				"operationId": "GetMessageDate",
				"responses": {
					"200": {
						"description": "Promise of messageDateResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/messageDateResponse"
								}
							}
						}
					}
				},
				"description": "EndPoint to retrieve a Message \"Good bye {name}\" and Date in json",
				"tags": [
					"goodByeController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/hello": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Promise of BasicResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponse"
								}
							}
						}
					}
				},
				"description": "EndPoint to retrieve a Message \"Hello {name}\" in json",
				"tags": [
					"helloController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Endpoint to retrieve the users in the collection \"user\" of DB",
				"tags": [
					"IUserController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"IUserController"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}